Guidler Pivot Location Revision
In an attempt to increase tension adjustment linearity, and improve the ability to print flexible materials, the guidler pivot has been re-located.  

The relocation allows for the driven filament gear to travel in a large arc which reduces the amount of clearance required around the gear (less place for sneaky squishy filament to buckle and work it's way out of the filament path.  

The pivot has also been moved to be closer to vertical with the axis of the driven gear which means that movement of the driven gear in and out (from adapting to gears of different specs) results in less vertical misalignment of the two gears which forces the filament of axis from the filament path.