Revised geometry for shuttle/main body/motor plate for new pivot
Since the proof-of-concept of the revised guidler pivot location seems to be yielding good results, now that I had some time I revised the geometry of the shuttle and all other affected components to have their swing-arc at the revised pivot point rather than the original.

Revised CW2 parts in this commit are *not* backwards compatible.

**Major Changes**
* redesigned `latch shuttle` from scratch
* adjusted `main body` and `motor plate` geometry for new shuttle swing-arc
* revised latch shuttle orientation tab/slot on `guidler a` and `guidler b`

**Minor Changes**
* corrected slicer issue on `guidler_b`