Guidler Pivot Relocation (#104)
* Guidler Pivot Location Revision

In an attempt to increase tension adjustment linearity, and improve the ability to print flexible materials, the guidler pivot has been re-located.  

The relocation allows for the driven filament gear to travel in a large arc which reduces the amount of clearance required around the gear (less place for sneaky squishy filament to buckle and work it's way out of the filament path.  

The pivot has also been moved to be closer to vertical with the axis of the driven gear which means that movement of the driven gear in and out (from adapting to gears of different specs) results in less vertical misalignment of the two gears which forces the filament of axis from the filament path.

* Revised geometry for shuttle/main body/motor plate for new pivot

Since the proof-of-concept of the revised guidler pivot location seems to be yielding good results, now that I had some time I revised the geometry of the shuttle and all other affected components to have their swing-arc at the revised pivot point rather than the original.

Revised CW2 parts in this commit are *not* backwards compatible.

**Major Changes**
* redesigned `latch shuttle` from scratch
* adjusted `main body` and `motor plate` geometry for new shuttle swing-arc
* revised latch shuttle orientation tab/slot on `guidler a` and `guidler b`

**Minor Changes**
* corrected slicer issue on `guidler_b`

* Shortened Latch Shuttle

Shortened new latch shuttle to increase guidler arm travel.

* Renamed Pre-Beta Files for Release

* Updated Changelog